+++ src/bin/ipa_bench/gen_events.rs
